mindmapper/backend/requirements.txt
fastapi
uvicorn
watchdog
python-dotenv
google-generativeai
pydantic
